% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_2d}
\alias{desc_2d}
\title{Compute all 2D molecular descriptors}
\usage{
desc_2d(mols)
}
\arguments{
\item{mols}{A \code{tidymol} object.}
}
\description{
Compute all 2D molecular descriptors
}
\examples{
\dontrun{
mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

desc_2d(mol)
desc_2d(mols)}
}
