% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc_3d}
\alias{desc_3d}
\title{Compute all 3D molecular descriptors.}
\usage{
desc_3d(mols)
}
\arguments{
\item{mols}{A \code{tidymol} object.}
}
\description{
Compute all 3D descriptors available in RDKit.
}
\note{
Since there is some randomization in the conformer generation
process, there would be some small numerical differences in the
calculated descriptors in each different run.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

desc_3d(mol)
desc_3d(mols)
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
