% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_atompair}
\alias{fp_atompair}
\title{Compute atom pairs fingerprints}
\usage{
fp_atompair(mols)
}
\arguments{
\item{mols}{A \code{tidymol} object.}
}
\description{
Compute atom pairs fingerprints
}
\note{
The vectorized version of this fingerprint has a dimensionality
of millions, which is usually considered too high for direct
statistical modeling. Therefore we did not include the option
to return the explicit fingerprint matrix here.
}
\examples{
\dontshow{if (is_installed_rdkit()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_atompair(mol)
fp_atompair(mols)
\dontshow{\}) # examplesIf}
}
