% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_maccs}
\alias{fp_maccs}
\title{Compute MACCS keys as molecular fingerprints}
\usage{
fp_maccs(mols, explicit = FALSE)
}
\arguments{
\item{mols}{A \code{tidymol} object.}

\item{explicit}{Return the fingerprints as a vector or matrix?
Default is \code{FALSE}.}
}
\description{
Compute MACCS keys as molecular fingerprints
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_maccs(mol)
fp_maccs(mols)

# get fingerprints matrix
fp_maccs(mol, explicit = TRUE)
fp_maccs(mols, explicit = TRUE)
}
}
