% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_morgan}
\alias{fp_morgan}
\title{Compute Morgan fingerprints (circular fingerprints).}
\usage{
fp_morgan(mols, radius = 2L, use_features = FALSE, explicit = FALSE)
}
\arguments{
\item{mols}{A \code{tidymol} object.}

\item{radius}{The radius parameter for Morgan fingerprints.
\code{radius = 2} is roughly equivalent to ECFP4 and FCFP4.}

\item{use_features}{Should we use the feature-based invariants,
similar to those used for the FCFP fingerprints? Defaults to \code{FALSE}.
This means only apply the default atom invariants, which use the
connectivity information similar to those used for the ECFP fingerprints.}

\item{explicit}{Return the fingerprints as a vector or matrix?
Default is \code{FALSE}.}
}
\description{
Compute Morgan fingerprints (circular fingerprints).
}
\examples{
\dontshow{if (is_installed_rdkit()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

# ECFP4
fp_morgan(mol)
fp_morgan(mols)

# get fingerprints matrix
fp_morgan(mol, explicit = TRUE)
fp_morgan(mols, explicit = TRUE)

# FCFP2
fp_morgan(mol, radius = 1L, use_features = TRUE)
fp_morgan(mols, radius = 1L, use_features = TRUE)

# get fingerprints matrix
fp_morgan(mol, radius = 1L, use_features = TRUE, explicit = TRUE)
fp_morgan(mols, radius = 1L, use_features = TRUE, explicit = TRUE)
\dontshow{\}) # examplesIf}
}
