% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_pattern}
\alias{fp_pattern}
\title{Compute pattern fingerprints, a topological fingerprint optimized for substructure screening}
\usage{
fp_pattern(mols, explicit = FALSE)
}
\arguments{
\item{mols}{A \code{tidymol} object.}

\item{explicit}{Return the fingerprints as a vector or matrix?
Default is \code{FALSE}.}
}
\description{
Compute pattern fingerprints, a topological fingerprint optimized for substructure screening
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_pattern(mol)
fp_pattern(mols)

fp_pattern(mol, explicit = TRUE)
fp_pattern(mols, explicit = TRUE)
}
}
