% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_pharm2d}
\alias{fp_pharm2d}
\title{Compute 2D pharmacophore fingerprints.}
\usage{
fp_pharm2d(mols, type = c("default", "gobbi"))
}
\arguments{
\item{mols}{A \code{tidymol} object.}

\item{type}{Type of the fingerprint. \code{"default"} uses
the chemical features defined by RDKit, \code{"gobbi"} uses
the set of chemical features feature defined by Gobbi and Poppinger (1998).}
}
\description{
Compute 2D pharmacophore fingerprints.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_pharm2d(mol)
fp_pharm2d(mols)

fp_pharm2d(mol, type = "gobbi")
fp_pharm2d(mols, type = "gobbi")
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
