% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_rdkit}
\alias{fp_rdkit}
\title{Compute RDKit fingerprints.}
\usage{
fp_rdkit(mols, explicit = FALSE)
}
\arguments{
\item{mols}{A \code{tidymol} object.}

\item{explicit}{Return the fingerprints as a vector or matrix?
Default is \code{FALSE}.}
}
\description{
Compute the Daylight-like fingerprint based on
hashing molecular subgraphs defined by RDKit.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_rdkit(mol)
fp_rdkit(mols)

# get fingerprints matrix
fp_rdkit(mol, explicit = TRUE)
fp_rdkit(mols, explicit = TRUE)
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
