% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_torsion}
\alias{fp_torsion}
\title{Compute topological torsions fingerprints.}
\usage{
fp_torsion(mols)
}
\arguments{
\item{mols}{A \code{tidymol} object.}
}
\description{
Compute topological torsions fingerprints.
}
\note{
The vectorized version of this fingerprint has a dimensionality
of millions, which is usually considered too high for direct
statistical modeling. Therefore we did not include the option
to return the explicit fingerprint matrix here.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_torsion(mol)
fp_torsion(mols)
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
