% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_rdkit}
\alias{install_rdkit}
\title{Install rdkit and its dependencies}
\usage{
install_rdkit(
  ...,
  envname = "r-rdkit",
  new_env = identical(envname, "r-rdkit")
)
}
\arguments{
\item{...}{Other arguments passed to \code{\link[reticulate:py_install]{reticulate::py_install()}}.}

\item{envname}{The name or full path of the environment in which
rdkit is installed. Default is \code{r-rdkit}.}

\item{new_env}{Logical. If \code{TRUE}, the specified Python environment
will be deleted and recreated if it already exists.
Defaults to \code{TRUE} only when using the default environment name.}
}
\value{
Invisibly returns \code{NULL}. Primarily used for its side effect
of installing the Python package in the specified environment.
}
\description{
Install rdkit and its dependencies
}
\examples{
\dontshow{if (is_installed_rdkit()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
install_rdkit()
\dontshow{\}) # examplesIf}
}
