% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_installed.R
\name{is_installed_rdkit}
\alias{is_installed_rdkit}
\title{Is rdkit installed?}
\usage{
is_installed_rdkit()
}
\value{
\code{TRUE} if installed, \code{FALSE} if not.
}
\description{
Is rdkit installed?
}
\examples{
is_installed_rdkit()
}
