% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_sdf}
\alias{parse_sdf}
\title{Parse molecules from SDF/MOL text block.}
\usage{
parse_sdf(string)
}
\arguments{
\item{string}{A SDF/MOL text block string.}
}
\description{
Parse molecules from SDF/MOL text block.
}
\examples{
\dontrun{

str_single <- readr::read_file(tidychem_example("single.mol"))
str_multiple <- readr::read_file(tidychem_example("5ht3ligs.sdf"))

parse_sdf(str_single)
parse_sdf(str_multiple)
}
}
