% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_smiles}
\alias{parse_smiles}
\title{Parse molecules from SMILES strings.}
\usage{
parse_smiles(string)
}
\arguments{
\item{string}{A vector of SMILES string(s).}
}
\description{
Parse molecules from a vector of SMILES strings.
}
\examples{
\dontrun{

parse_smiles("Cc1ccccc1")
parse_smiles(c("Cc1ccccc1", "CCN(CC)CC"))
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
