% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidydesc}
\alias{print.tidydesc}
\title{Print molecular descriptors object.}
\usage{
\method{print}{tidydesc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidydesc}.}

\item{...}{Additional parameters for \code{\link[=print]{print()}} (not used).}
}
\description{
Print molecular descriptors object.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

desc_2d(mol)
desc_2d(mols)
}
}
