% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidyfps}
\alias{print.tidyfps}
\title{Print molecular fingerprints object.}
\usage{
\method{print}{tidyfps}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidyfps}.}

\item{...}{Additional parameters for \code{\link{print}} (not used).}
}
\description{
Print molecular fingerprints object.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_topo(mol)
fp_topo(mols)

fp_topo(mol, explicit = TRUE)
fp_topo(mols, explicit = TRUE)
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
