% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdkit_version}
\alias{rdkit_version}
\title{Get RDKit version.}
\usage{
rdkit_version()
}
\description{
Return the version number of the installed RDKit.
}
\examples{
\dontrun{

rdkit_version()
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
