% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_smiles}
\alias{read_smiles}
\title{Read molecules from a SMILES file.}
\usage{
read_smiles(file)
}
\arguments{
\item{file}{Path to the SMILES file.}
}
\description{
Read molecules from a SMILES file.
}
\examples{
\dontshow{if (is_installed_rdkit()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
read_smiles(tidychem_example("smi-single.smi"))
read_smiles(tidychem_example("smi-multiple.smi"))
read_smiles(tidychem_example("smi-single-noname.smi"))
read_smiles(tidychem_example("smi-multiple-noname.smi"))
\dontshow{\}) # examplesIf}
}
