% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidychem_example}
\alias{tidychem_example}
\title{Generate path to the example data.}
\usage{
tidychem_example(file)
}
\arguments{
\item{file}{Filename of the example dataset in the tidychem package,
under the \verb{inst/extdata/} directory.}
}
\description{
Generate path to the example data.
}
