% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fp.R
\name{fp_erg}
\alias{fp_erg}
\title{Compute extended reduced graphs fingerprints.}
\usage{
fp_erg(mols)
}
\arguments{
\item{mols}{A \code{tidymol} object.}
}
\description{
Compute extended reduced graphs fingerprints
derived from the ErG fingerprint.
}
\examples{
\dontrun{

mol <- parse_smiles("Cc1ccccc1")
mols <- read_smiles(tidychem_example("smi-multiple.smi"))

fp_erg(mol)
fp_erg(mols)
}
}
\author{
Nan Xiao <\url{https://nanx.me}>
}
