% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tidymol}
\alias{print.tidymol}
\title{Print parsed molecules object.}
\usage{
\method{print}{tidymol}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tidymol}.}

\item{...}{Additional parameters for \code{\link{print}} (not used).}
}
\description{
Print parsed molecules object.
}
\examples{
\dontrun{

parse_smiles("Cc1ccccc1")
parse_smiles(c("Cc1ccccc1", "CCN(CC)CC"))
}
}
