% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rdkit}
\alias{rdkit}
\title{Global reference to RDKit}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
rdkit
}
\description{
Global reference to RDKit which will be initialized in \code{.onLoad}.
}
\keyword{datasets}
