% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rdkit_version}
\alias{rdkit_version}
\title{Get RDKit version.}
\usage{
rdkit_version()
}
\description{
Get RDKit version.
}
\examples{
\dontshow{if (is_installed_rdkit()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rdkit_version()
\dontshow{\}) # examplesIf}
}
