% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sdf}
\alias{read_sdf}
\title{Read molecules from a SDF/MOL file.}
\usage{
read_sdf(file)
}
\arguments{
\item{file}{Path to the SDF or MOL file.}
}
\description{
Read molecules from a SDF/MOL file.
}
\examples{
\dontrun{

read_sdf(tidychem_example("single.mol"))
read_sdf(tidychem_example("5ht3ligs.sdf"))
}
}
