% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Calculate fingerprint similarity.}
\usage{
sim(
  fps1 = NULL,
  fps2 = NULL,
  metric = c("tanimoto", "dice", "cosine", "sokal", "russel", "rogotgoldberg",
    "allbit", "kulczynski", "mcconnaughey", "asymmetric", "braunblanquet")
)
}
\arguments{
\item{fps1}{A \code{tidyfps} object.}

\item{fps2}{A \code{tidyfps} object (optional).}
}
\description{
Calculate fingerprint similarity.
}
\examples{
\dontrun{

mols1 <- parse_smiles(c("Cc1ccccc1", "CCN(CC)CC"))
mols2 <- parse_smiles(c("Cc1ccccc1", "CCN(CC)CC", "CCC"))
fps1 <- fp_maccs(mols1)
fps2 <- fp_maccs(mols2)

sim(fps1, metric = "tanimoto")
sim(fps1, fps2, metric = "tanimoto")

# or use pipes
mols1 \%>\% fp_maccs \%>\% sim("tanimoto")
(mols1 \%>\% fp_maccs) \%>\% sim(mols2 \%>\% fp_maccs, "tanimoto")
}
}
