% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychem-package.R
\docType{package}
\name{tidychem-package}
\alias{tidychem-package}
\title{Tidy and Featurize Chemical Data using 'RDKit'}
\description{
Read, preprocess, and featurize chemical data using 'RDKit'.
Offers a digested, functional, tidyverse-style interface.
}
\details{
This package also offers a lightweight R interface for accessing RDKit
via the \href{http://www.rdkit.org/docs/api/}{RDKit Python API} and
the \code{reticulate} package.

\tabular{ll}{
Package: \tab tidychem\cr
Type: \tab Package\cr
License: \tab BSD_3_clause}
}
