% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tidychem-package}
\alias{tidychem}
\alias{tidychem-package}
\title{tidychem: Tidy Chemical Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Read, preprocess, and featurize chemical data using 'RDKit'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/tidychem/}
  \item \url{https://github.com/nanxstats/tidychem/}
  \item Report bugs at \url{https://github.com/nanxstats/tidychem/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

}
\keyword{internal}
